requirement.txt
numpy>=1.26.0
pillow>=9.5.0
easyocr>=1.4.1